Move methods for getting the Reference out of the controller
The Reference is mostly drawn from attributes we have in the Request model, so it makes sense to keep it there.
