+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/v2/IdentifierResource.java
@@ -31,7 +31,7 @@ public class IdentifierResource extends IIIF2Resource {
     @Override
     public void doGET() {
         final Reference newRef = new Reference(
-                getPublicRootReference() +
+                getRequest().getPublicRootReference() +
                 Route.IIIF_2_PATH +
                 "/" + getPublicIdentifier() +
                 "/info.json");
