+++ src/test/java/edu/illinois/library/cantaloupe/resource/AbstractResourceTest.java
@@ -99,92 +99,6 @@ public class AbstractResourceTest extends BaseTest {
         assertTrue(types.isEmpty());
     }
 
-    /**
-     * Tests behavior of {@link AbstractResource#getPublicReference()} when
-     * using {@link Key#BASE_URI}.
-     */
-    @Test
-    void testGetPublicReferenceUsingConfiguration() {
-        final String baseURI = "http://example.net/base";
-        Configuration.getInstance().setProperty(Key.BASE_URI, baseURI);
-
-        MockHttpServletRequest servletRequest =
-                (MockHttpServletRequest) instance.getRequest().getServletRequest();
-        servletRequest.setContextPath("/base");
-        servletRequest.setRequestURL("http://example.org/base/llamas");
-
-        Reference ref = instance.getPublicReference();
-        assertEquals(baseURI + "/llamas", ref.toString());
-    }
-
-    /**
-     * Tests behavior of {@link AbstractResource#getPublicReference()} when
-     * using {@literal X-Forwarded} headers.
-     *
-     * This isn't a thorough test of every possible header/URI combination.
-     * See {@link Reference#applyProxyHeaders(Headers)} for those.
-     */
-    @Test
-    void testGetPublicReferenceUsingXForwardedHeaders() {
-        MockHttpServletRequest servletRequest =
-                (MockHttpServletRequest) instance.getRequest().getServletRequest();
-        servletRequest.setContextPath("");
-        servletRequest.setRequestURL("http://bogus/cats");
-
-        Headers headers = instance.getRequest().getHeaders();
-        headers.set("X-Forwarded-Proto", "HTTP");
-        headers.set("X-Forwarded-Host", "example.org");
-        headers.set("X-Forwarded-Port", "80");
-        headers.set("X-Forwarded-Path", "/");
-        Reference ref = instance.getPublicReference();
-        assertEquals("http://example.org/cats", ref.toString());
-    }
-
-    /**
-     * Tests behavior of {@link AbstractResource#getPublicReference()} when
-     * using neither {@link Key#BASE_URI} nor {@literal X-Forwarded} headers.
-     */
-    @Test
-    void testGetPublicReferenceFallsBackToHTTPRequest() {
-        String resourceURI = "http://example.net/cats/dogs";
-
-        MockHttpServletRequest servletRequest =
-                (MockHttpServletRequest) instance.getRequest().getServletRequest();
-        servletRequest.setContextPath("/cats");
-        servletRequest.setRequestURL(resourceURI);
-        Reference ref = instance.getPublicReference();
-        assertEquals(resourceURI, ref.toString());
-    }
-
-    /**
-     * Tests behavior of {@link AbstractResource#getPublicReference()} when
-     * using neither {@link Key#BASE_URI} nor {@literal X-Forwarded} headers.
-     */
-    @Test
-    void testGetPublicReferenceFallsBackToHTTPSRequest() {
-        String resourceURI = "https://example.net/cats/dogs";
-
-        MockHttpServletRequest servletRequest =
-                (MockHttpServletRequest) instance.getRequest().getServletRequest();
-        servletRequest.setContextPath("/cats");
-        servletRequest.setRequestURL(resourceURI);
-        Reference ref = instance.getPublicReference();
-        assertEquals(resourceURI, ref.toString());
-    }
-
-    @Test
-    void testGetPublicReferenceOmitsQuery() {
-        String resourceURI = "https://example.net/cats/dogs?arg=value";
-        String expected = "https://example.net/cats/dogs";
-
-        MockHttpServletRequest servletRequest =
-                (MockHttpServletRequest) instance.getRequest().getServletRequest();
-        servletRequest.setContextPath("/cats");
-        servletRequest.setRequestURL(resourceURI);
-        Reference ref = instance.getPublicReference();
-        assertEquals(expected, ref.toString());
-    }
-
     /* getRepresentationDisposition() */
 
     @Test
